+++ Journey/ClientApp/src/app/app.component.html
@@ -1,6 +1,3 @@
-<body>
-  <app-nav-bar></app-nav-bar>
-  <div class="container">
+
     <router-outlet></router-outlet>
-  </div>
-</body>
+
