+++ Journey/ClientApp/src/app/app.module.ts
@@ -5,8 +5,6 @@ import { HttpClientModule } from '@angular/common/http';
 import { AppComponent } from './app.component';
 import { AppRoutingModule } from './app-routing.module';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { HeaderModule } from './shared/header/header.module';
-import { NavBarComponent } from './shared/header/components/nav-bar/nav-bar.component';
 
 @NgModule({
   declarations: [
@@ -14,11 +12,8 @@ import { NavBarComponent } from './shared/header/components/nav-bar/nav-bar.comp
   ],
   imports: [
     BrowserModule.withServerTransition({ appId: 'ng-cli-universal' }),
-    HttpClientModule,
-    FormsModule,
     AppRoutingModule,
-    BrowserAnimationsModule,
-    HeaderModule
+    BrowserAnimationsModule
   ],
   providers: [],
   bootstrap: [AppComponent]
