+++ Journey/ClientApp/src/app/destination/destination/destination.component.css
@@ -1,0 +1,9 @@
+html,body{
+    height: 100% !important; 
+    background-image: url(../../../assets/exDestination.png);
+    background-size: cover !important;
+    background-attachment: fixed;
+    min-height: 100vh;
+    height: auto;
+    margin: 0;
+}
