+++ Journey/ClientApp/src/app/destination/destination/destination.component.html
@@ -1,0 +1,10 @@
+<html>
+    <head>
+
+    </head>
+    <body>
+        <app-nav-bar></app-nav-bar>
+        <app-short-description></app-short-description>
+    </body>
+</html>
+
