+++ Journey/ClientApp/src/app/destination/destination/destination.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-destination',
+  templateUrl: './destination.component.html',
+  styleUrls: ['./destination.component.css']
+})
+export class DestinationComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
