+++ Journey/ClientApp/src/app/destination/short-description/short-description.component.html
@@ -1,1 +1,8 @@
-<p>short-description works!</p>
+<html>
+    <head>
+
+    </head>
+    <body>
+        Test
+    </body>
+</html>
