+++ Journey/ClientApp/src/app/shared/components/nav-bar/nav-bar.component.html
@@ -1,0 +1,4 @@
+<mat-toolbar>
+    <span><img src="../../../../assets/journey_icon.png"></span>
+    <span>journey</span>
+</mat-toolbar>
