--- Journey/ClientApp/src/app/shared/header/components/nav-bar/nav-bar.component.html
@@ -1,4 +1,0 @@
-<mat-toolbar>
-    <span>Custom Toolbar</span>
-    <span>journey</span>
-</mat-toolbar>
