+++ Journey/ClientApp/src/styles.css
@@ -1,37 +1,4 @@
 /* You can add global styles to this file, and also import other style files */
 
-/* Provide sufficient contrast against white background */
-a {
-  color: #0366d6;
-}
-
-code {
-  color: #e01a76;
-}
-
-.btn-primary {
-  color: #fff;
-  background-color: #1b6ec2;
-  border-color: #1861ac;
-}
-
-*
-{
-  padding: 0;
-  margin: 0;
-} 
-
-body{
-  height: 100% !important; 
-  background-image: linear-gradient(rgba(0,0,0,0.6), rgba(0,0,0,0.6)), url(assets/img.jpg);
-  background-size: cover !important;
-  background-position: center;
-  background-attachment: fixed;
-}
-h1 {
-  color: #ffffff;
-  font-family: Arial, Helvetica, sans-serif;
-  font-size: 250%;
-}
 html, body { height: 100%; }
-body { margin: 0; font-family: Roboto, "Helvetica Neue", sans-serif; }
+body { margin: 0; font-family: Roboto, "Helvetica Neue", sans-serif; color: white; }
